Move resources.py
