+++ bot/resources.py
